refactor: 참가자 공통 로직 추상화
