+++ README.md
@@ -48,18 +48,22 @@
 - [x] 카드를 한장 뽑는다.
     - [x] 더이상 뽑을 카드가 없는 경우 예외가 발생한다.
 
-### 플레이어
+### 참가자 공통 규칙
 
-- [x] 이름과 패를 가진다.
+- [x] 패를 가진다.
 - [x] 카드를 한장 받는다.
 - [x] 가진 패의 합을 반환한다.
+
+### 플레이어
+
+- [x] 참가자 공통 규칙을 지킨다.
+- [x] 이름을 가진다.
 
 ### 딜러
 
-- [x] 패와 덱을 가진다.
+- [x] 참가자 공통 규칙을 지킨다.
+- [x] 덱을 가진다.
 - [x] 덱에서 카드를 한장 뽑는다.
-- [x] 카드를 한장 받는다.
-- [x] 가진 패의 합을 반환한다.
 
 ## 실행 예시
